backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
pillow==10.0.1
opencv-python==4.8.0.76
tensorflow==2.15.0
scikit-image==0.21.0
transformers==4.34.0
torch==2.0.1
